# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dega server API",
        "title": "Dega API",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:7789",
    "basePath": "/",
    "paths": {
...
        "/core/menus": {
            "get": {
                "description": "Get all menus",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "Show all menus",
                "operationId": "get-all-menus",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "limit per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "page number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/menu.paging"
                        }
                    }
                }
            },
            "post": {
                "description": "Create menu",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "Create menu",
                "operationId": "add-menu",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Menu Object",
                        "name": "Menu",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/menu.menu"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Menu"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/core/pages/{page_id}": {
            "get": {
                "description": "Get page by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Page"
                ],
                "summary": "Show a page by id",
                "operationId": "get-page-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Page ID",
                        "name": "page_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/page.pageData"
                        }
                    }
                }
            },
            "put": {
                "description": "Update page by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Page"
                ],
                "summary": "Update a page by id",
                "operationId": "update-page-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Page ID",
                        "name": "page_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Page",
                        "name": "Page",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/page.page"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/page.pageData"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete page by ID",
                "tags": [
                    "Page"
                ],
                "summary": "Delete a page",
                "operationId": "delete-page-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Page ID",
                        "name": "page_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
...
        "/core/permissions/organisations/{permission_id}": {
            "put": {
                "description": "Update Organisation permission by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organisation_Permissions"
                ],
                "summary": "Update a Organisation permission by id",
                "operationId": "update-org-permission-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Permission ID",
                        "name": "permission_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Permission Body",
                        "name": "Permission",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/organisation.organisationPermission"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.OrganisationPermission"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Organisation permission by ID",
                "tags": [
                    "Organisation_Permissions"
                ],
                "summary": "Delete a Organisation permission",
                "operationId": "delete-org-permission-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Permission ID",
                        "name": "permission_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/core/permissions/spaces": {
            "get": {
                "description": "Get all Space permissions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Space_Permissions"
                ],
                "summary": "Show all Space permissions",
                "operationId": "get-all-space-permissions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Query",
                        "name": "q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github.com_factly_dega-server_service_core_action_permissions_space.paging"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create Space permission",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Space_Permissions"
                ],
                "summary": "Create Space permission",
                "operationId": "add-space-permission",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Permission Object",
                        "name": "Permission",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/space.spacePermission"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.SpacePermission"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github.com_factly_dega-server_service_core_action_permissions_space.paging": {
            "type": "object",
            "properties": {
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github.com_factly_dega-server_service_core_action_permissions_space.spaceWithPermissions"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "menu.menu": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "menu": {
                    "type": "string"
                },
                "meta_fields": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "menu.paging": {
            "type": "object",
            "properties": {
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Menu"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "model.Format": {
            "type": "object",
            "required": [
                "name",
                "slug"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "meta_fields": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "space": {
                    "$ref": "#/definitions/model.Space"
                },
                "space_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                }
            }
        },
...
        "model.Menu": {
            "type": "object",
            "required": [
                "name",
                "slug"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "menu": {
                    "type": "string"
                },
                "meta_fields": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "space": {
                    "$ref": "#/definitions/model.Space"
                },
                "space_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                }
            }
        },
        "model.OrganisationPermission": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "organisation_id": {
                    "type": "integer"
                },
                "spaces": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                }
            }
        },
...
        "model.SpacePermission": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "episodes": {
                    "type": "integer"
                },
                "fact_check": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "media": {
                    "type": "integer"
                },
                "podcast": {
                    "type": "boolean"
                },
                "posts": {
                    "type": "integer"
                },
                "space": {
                    "$ref": "#/definitions/model.Space"
                },
                "space_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                }
            }
        },
...
        "organisation.organisationPermission": {
            "type": "object",
            "required": [
                "organisation_id"
            ],
            "properties": {
                "organisation_id": {
                    "type": "integer"
                },
                "spaces": {
                    "type": "integer"
                }
            }
        },
...
        "page.page": {
            "type": "object",
            "required": [
                "format_id",
                "title"
            ],
            "properties": {
                "author_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "category_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "description": {
                    "type": "string"
                },
                "excerpt": {
                    "type": "string"
                },
                "featured_medium_id": {
                    "type": "integer"
                },
                "format_id": {
                    "type": "integer"
                },
                "is_featured": {
                    "type": "boolean"
                },
                "is_highlighted": {
                    "type": "boolean"
                },
                "is_sticky": {
                    "type": "boolean"
                },
                "published_date": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "space_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "subtitle": {
                    "type": "string"
                },
                "tag_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "page.pageData": {
            "type": "object",
            "properties": {
                "authors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Author"
                    }
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Category"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "excerpt": {
                    "type": "string"
                },
                "featured_medium_id": {
                    "type": "integer"
                },
                "footer_code": {
                    "type": "string"
                },
                "format": {
                    "$ref": "#/definitions/model.Format"
                },
                "format_id": {
                    "type": "integer"
                },
                "header_code": {
                    "type": "string"
                },
                "html_description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_featured": {
                    "type": "boolean"
                },
                "is_highlighted": {
                    "type": "boolean"
                },
                "is_page": {
                    "type": "boolean"
                },
                "is_sticky": {
                    "type": "boolean"
                },
                "medium": {
                    "$ref": "#/definitions/model.Medium"
                },
                "meta": {
                    "type": "string"
                },
                "meta_fields": {
                    "type": "string"
                },
                "published_date": {
                    "type": "string"
                },
                "schemas": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "space": {
                    "$ref": "#/definitions/model.Space"
                },
                "space_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "subtitle": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Tag"
                    }
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                }
            }
        },
...
        "space.spacePermission": {
            "type": "object",
            "required": [
                "space_id"
            ],
            "properties": {
                "episodes": {
                    "type": "integer"
                },
                "fact_check": {
                    "type": "boolean"
                },
                "media": {
                    "type": "integer"
                },
                "podcast": {
                    "type": "boolean"
                },
                "posts": {
                    "type": "integer"
                },
                "space_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}